==> 3_formula1_formula2.ipynb <==
import pandas as pd
from datetime import datetime

sorted_merge_visits = pd.read_csv('D://A_thesis/gowalla/output/sorted_merge_visits.csv')

sorted_merge_visits

location_density = sorted_merge_visits["visit_user_from_loc"]/sorted_merge_visits["total_visit"]

# location_density

formula1_df = sorted_merge_visits.assign(location_density=location_density)
formula1_df

formula1_df.to_csv('D://A_thesis/gowalla/output/formula1_df.csv')

formula1_df['location_density'].min()

**************************************************************************************
**************************************************************************************
**************************************************************************************

dataset_final = pd.read_csv('D://A_thesis/gowalla/dataset_final_with_date_day_ampm.csv')

dataset_final

dataset_final['checkin_date'] = pd.to_datetime(dataset_final['check-in time']).dt.date
dataset_final['checkin_time'] = pd.to_datetime(dataset_final['check-in time']).dt.strftime('%I:%M %p')

dataset_final = dataset_final.sort_values(by=['userid','checkin_date'],ascending = [True, True])
dataset_final

max_date = dataset_final.groupby(['userid'])['checkin_date'].max()
min_date = dataset_final.groupby(['userid'])['checkin_date'].min()

lifetime_user = pd.merge(max_date, min_date, on='userid')

lifetime_user.rename(columns={"checkin_date_x": "last time", "checkin_date_y": "first time"}, inplace = True)
lifetime_user.reset_index(inplace=True)

lifetime_user

lifetime = lifetime_user['last time'] - lifetime_user['first time']
# lifetime
lifetime_user['lifetime'] = lifetime

lifetime_user

lifetime_user.to_csv('D://A_thesis/gowalla/out_random/lifetime_user.csv')

dataset_final.to_csv('D://A_thesis/gowalla/out_random/dataset_final_before_drop.csv')

# dataset_final = dataset_final.drop(columns=['Unnamed: 0', 'latitude', 'longitude'])
# del dataset_final['Unnamed: 0']

# dataset_final

last_visit = dataset_final.groupby(['userid', 'location_id'])['checkin_date'].max()
# last_visit
first_visit = dataset_final.groupby(['userid', 'location_id'])['checkin_date'].min()
# first_visit

merge_visit = pd.merge(last_visit, first_visit, on=["userid","location_id"])
merge_visit

merge_visit.rename(columns={"checkin_date_x": "last_visit_LinU", "checkin_date_y": "first_visit_LinU"}, inplace = True)
merge_visit.reset_index(inplace=True)

merge_visit

difference_visit = merge_visit['last_visit_LinU'] - merge_visit['first_visit_LinU']
merge_visit['difference_visit'] = difference_visit

merge_visit

merge_visit.to_csv('D://A_thesis/gowalla/out_random/merge_visit.csv')

# merge_visit["difference_visit"] = merge_visit["difference_visit"].astype(int)
# merge_visit["difference_visit"] = merge_visit["difference_visit"].str.replace(" days","")

formula2_df = pd.read_csv('D://A_thesis/gowalla/out_random/lifetime_user.csv')

formula2_df

formula2_df["lifetime"] = formula2_df["lifetime"].str.replace(" 00:00:00.000000000","")

# formula2_df["lifetime"] = formula2_df["lifetime"].astype(int)

formula2_df['total_visit_lifetime'] = merge_visit['difference_visit'] /formula2_df['lifetime']

formula2_df

merge_visit_and_formula2_df = pd.merge(formula2_df, merge_visit, on=['userid'])

merge_visit_and_formula2_df

del merge_visit_and_formula2_df['Unnamed: 0']

merge_f1_f2 = pd.merge(formula1_df, merge_visit_and_formula2_df, on=['userid','location_id'])
merge_f1_f2

merge_f1_f2.drop(columns=['Unnamed: 0'])

merge_f1_f2.to_csv('D://A_thesis/gowalla/out_random/merge_f1_f2.csv')

# f1_f2 = pd.read_csv('D://merge_f1_f2.csv')

***************************************************************************************************
***************************************************************************************************
***************************************************************************************************
***************************************************************************************************

merge_f1_f2[merge_f1_f2['visit_user_from_loc'] == 1].count()

merge_f1_f2[merge_f1_f2['visit_user_from_loc'] > 1].count()

merge_f1_f2_delvisit1 = merge_f1_f2.loc[formula1_df['visit_user_from_loc'] > 1 ]
merge_f1_f2_delvisit1

del merge_f1_f2_delvisit1['index']

merge_f1_f2_delvisit1.to_csv('D://A_thesis/gowalla/out_random/merge_f1_f2_delvisit1.csv')

min_density = merge_f1_f2_delvisit1['location_density'].min()
min_density

min_visit = merge_f1_f2_delvisit1['total_visit_lifetime'].min()
min_visit

merge_f1_f2_delvisit1[merge_f1_f2_delvisit1['total_visit_lifetime'] == 0.0 ].count()
