==> 4_calculate SILU.ipynb <==
import pandas as pd
from datetime import datetime

dataset_final = pd.read_csv('D://A_thesis/gowalla/dataset_final_with_date_day_ampm.csv')

dataset_final

dataset_final.drop(columns = {'Unnamed: 0' , 'Unnamed: 0.1'})

pre = pd.DataFrame()

total_info_cat = pd.DataFrame()

total_info_cat['visit_user_from_cat'] = dataset_final.groupby(['userid','category'])['category'].count()

pre0 = dataset_final.loc[dataset_final['IsWeekend_int'] == 0 ]
pre1 = dataset_final.loc[dataset_final['IsWeekend_int'] == 1 ]

total_info_cat['visit_user_cat_weekday'] = pre0.groupby(['userid','category'])['IsWeekend_int'].count()
total_info_cat['visit_user_cat_weekend'] = pre1.groupby(['userid', 'category'])['IsWeekend_int'].count()

pre2 = dataset_final.loc[dataset_final['AM_PM'] == 'AM' ]
pre3 = dataset_final.loc[dataset_final['AM_PM'] == 'PM' ]
total_info_cat['visit_user_cat_AM'] = pre2.groupby(['userid','category'])['AM_PM'].count()
total_info_cat['visit_user_cat_PM'] = pre3.groupby(['userid','category'])['AM_PM'].count()

total_info_cat['visit_user_cat_weekday'].fillna(0 , inplace = True)
total_info_cat['visit_user_cat_weekend'].fillna(0 , inplace = True)
total_info_cat["visit_user_cat_weekday"] = total_info_cat["visit_user_cat_weekday"].astype(int)
total_info_cat["visit_user_cat_weekend"] = total_info_cat["visit_user_cat_weekend"].astype(int)

total_info_cat

total_info_cat['visit_user_cat_AM'].fillna(0 , inplace = True)
total_info_cat['visit_user_cat_PM'].fillna(0 , inplace = True)
total_info_cat["visit_user_cat_AM"] = total_info_cat["visit_user_cat_AM"].astype(int)
total_info_cat["visit_user_cat_PM"] = total_info_cat["visit_user_cat_PM"].astype(int)

total_info_cat

total_info_cat.reset_index(inplace=True)

total_info_cat.to_csv('D://A_thesis/gowalla/out_random/total_info_cat.csv')

total_info_cat

total_info_cat['visit_user_cat_PM'].fillna(0 , inplace = True)

total_info_cat["visit_user_cat_AM"] = total_info_cat["visit_user_cat_AM"].astype(int)
total_info_cat["visit_user_cat_PM"] = total_info_cat["visit_user_cat_PM"].astype(int)

# total_info_cat

total_info_cat.to_csv('D://A_thesis/gowalla/out_random/total_info_cat.csv')

del total_info_cat2

total_info_cat2 = pd.DataFrame()

pre0 = dataset_final.loc[dataset_final['IsWeekend_int'] == 0 ]
pre1 = dataset_final.loc[dataset_final['IsWeekend_int'] == 1 ]

total_info_cat2['visit_user_cat_weekday'] = pre0.groupby(['userid','category'])['IsWeekend_int'].count()
total_info_cat2['visit_user_cat_weekend'] = pre1.groupby(['userid', 'category'])['IsWeekend_int'].count()

total_info_cat2.reset_index(inplace=True)

total_info_cat2['visit_user_cat_weekday'].fillna(0 , inplace = True)
total_info_cat2['visit_user_cat_weekend'].fillna(0 , inplace = True)

total_info_cat2["visit_user_cat_weekday"] = total_info_cat2["visit_user_cat_weekday"].astype(int)
total_info_cat2["visit_user_cat_weekend"] = total_info_cat2["visit_user_cat_weekend"].astype(int)

total_info_cat2

total_info_cat = pd.merge(total_info_cat2, total_info_cat, on =['userid', 'category'])

total_info_cat

total_info_cat.drop(columns = {'visit_user_cat_weekday_y' , 'visit_user_cat_weekend_y'})

total_info_cat.rename(columns = {'visit_user_cat_weekday_x':'visit_user_cat_weekday'  , 'visit_user_cat_weekend_x' :' visit_user_cat_weekend' })

# del total_info_cat['visit_user_cat_weekday_y']
# del total_info_cat['visit_user_cat_weekend_y']
# del total_info_cat['Unnamed: 0']

total_info_cat.to_csv('D://A_thesis/gowalla/out_random/total_info_cat.csv')

total_info_cat = pd.read_csv('D://A_thesis/gowalla/out_random/total_info_cat.csv')

total_info_cat

total_info_cat.rename(columns = {'visit_user_from_cat':'count_visit_user_from_cat'  , 'visit_user_cat_weekday' :' count_visit_user_cat_weekday' , 'visit_user_cat_weekend' : 'count_visit_user_cat_weekend', 'visit_user_cat_AM' : 'count_visit_user_cat_AM' , 'visit_user_cat_PM' : 'count_visit_user_cat_PM'})

total_info_cat.to_csv('D://A_thesis/gowalla/out_random/total_info_cat.csv')

***************************************************************************************************
***************************************************************************************************
***************************************************************************************************

user = dataset_final['userid'].groupby(dataset_final['userid']).unique()

random_user = dataset_final['userid'].unique()

del user_loc_lists
user_loc_lists = pd.DataFrame()

user_loc_lists['userid'] = random_user

user_loc_lists

user_loc_lists['user_loc_list'] = 1

allLocations =[]

merge_f1_f2_delvisit1 = pd.read_csv('D://A_thesis/gowalla/out_random/merge_f1_f2_delvisit1.csv')

merge_f1_f2_delvisit1

merge_f1_f2_delvisit1['location_density'] = merge_f1_f2_delvisit1['location_density'].round(decimals = 4)

merge_f1_f2_delvisit1

min_density = merge_f1_f2_delvisit1['location_density'].min()

min_density

merge_f1_f2_delvisit1["difference_visit"] = merge_f1_f2_delvisit1["difference_visit"].str.replace(" 00:00:00.000000000","")
merge_f1_f2_delvisit1["difference_visit"] = merge_f1_f2_delvisit1["difference_visit"].str.replace(" days","")
merge_f1_f2_delvisit1["lifetime"] = merge_f1_f2_delvisit1["lifetime"].str.replace(" days","")

merge_f1_f2_delvisit1["difference_visit"] = merge_f1_f2_delvisit1["difference_visit"].astype(int)
merge_f1_f2_delvisit1["lifetime"] = merge_f1_f2_delvisit1["lifetime"].astype(int)

merge_f1_f2_delvisit1['total_visit_lifetime'] = merge_f1_f2_delvisit1['difference_visit'] /merge_f1_f2_delvisit1['lifetime']

min_visit = merge_f1_f2_delvisit1['total_visit_lifetime'].min()

min_visit

mini_f1_f2 = pd.DataFrame()

mini_f1_f2['userid'] = merge_f1_f2_delvisit1['userid']
mini_f1_f2['location_id'] = merge_f1_f2_delvisit1['location_id']
mini_f1_f2['location_density'] = merge_f1_f2_delvisit1['location_density']
mini_f1_f2['total_visit_lifetime'] =merge_f1_f2_delvisit1['total_visit_lifetime'] 

mini_f1_f2

min_density = mini_f1_f2['location_density'].min()

min_density

min_visit = mini_f1_f2['total_visit_lifetime'].min()

min_density

user_loc_lists = pd.read_csv('D://A_thesis/gowalla/out_random/user_loc_lists.csv')

def user_loc(user) :
    Silu_user = []
    for i in range(len(mini_f1_f2)) :
        if mini_f1_f2['userid'][i] == user :
            if mini_f1_f2['location_density'][i] >= 0.0004 and mini_f1_f2['total_visit_lifetime'][i]>= 0.0 :
                Silu_user.append(mini_f1_f2['location_id'][i])
            
    return Silu_user  

# user_loc(1)

for i in range(len(user_loc_lists)) :
    user_loc_lists['user_loc_list'][i] =  user_loc(user_loc_lists['userid'][i])

user_loc_lists

user_loc_lists.to_csv('D://A_thesis/gowalla/out_random/user_loc_lists.csv')
user_loc_lists.to_csv('D://A_thesis/gowalla/out_random/user_loc_lists2.csv')
