==> 10_work with time and all_day.ipynb <==
import pandas as pd
from datetime import datetime

dataset_final = pd.read_csv('D://A_thesis/gowalla/out_random/dataset_final_before_drop.csv')

dataset_final = dataset_final.drop(columns = {'Unnamed: 0' , 'Unnamed: 0.1' , 'Unnamed: 0.1.1'})

dataset_final.dtypes

dataset_final['check-in time'] = dataset_final['check-in time'].map(lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S"))
dataset_final['time-24'] = pd.to_datetime(dataset_final['check-in time']).dt.strftime('%H:%M:%S')

            ***************************************************
            ***************************************************
            ***************************************************

# dataset_final['check-in time'] = dataset_final['check-in time'].map(lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S"))

dataset_final.dtypes

# dataset_final['time-24'] = datetime.strptime(dataset_final['check-in time'], "%H:%M:%S")
# dataset_final['time-24'] = dataset_final['time-24'].time()
# hour = dataset_final['time-24'].hour
# print("Hour : {}".format(hour))


# dataset_final['time-24'] = pd.to_datetime(dataset_final['check-in time']).dt.strftime('%H:%M %S')
# dataset_final['Hour-time'] = pd.to_datetime(dataset_final['check-in time']).dt.strftime('%H:%M %S').hour
# dataset_final['Hour-time'] = datetime.strptime(dataset_final['check-in time'],'%H:%M:%S').hour
# dataset_final['time-24'] = str(dataset_final['time-24'])
# hour = dataset_final['time-24'].hour
# print("Hour : {}".format(hour))

dataset_final['hour'] = dataset_final['check-in time'].apply(lambda x: x.hour)

# dataset_final.drop(columns = {'Unnamed: 0' , 'Unnamed: 0.1' , 'Unnamed: 0.1.1'})

            ***************************************************
            ***************************************************
            ***************************************************

dataset_final["DayOfWeek"] = dataset_final["check-in time"].dt.weekday

dataset_final["dayName"] = pd.to_datetime(dataset_final['check-in time']).dt.strftime("%A")

dataset_final["dayName"]

# dataset_final = dataset_final.drop(columns = {'Unnamed: 0' , 'Unnamed: 0.1' , 'Unnamed: 0.1.1'})

dataset_final

dataset_final['hour'].min()

# def get_part_of_day(h):
#     return (
#         "morning"
#         if 5 <= h <= 11
#         else "afternoon"
#         if 12 <= h <= 17
#         else "evening"
#         if 18 <= h <= 22
#         else "night"
#     )

# get_part_of_day(19)

dataset_final['status_day'] = 1

# for i in range(len(dataset_final)) :
#     dataset_final['status_day'][i] = get_part_of_day(dataset_final['hour'][i])

# dataset_final_new = dataset_final[:1000000]

# def status_of_day(h):
#     for i in range(len(dataset_final_new)) :
#         if dataset_final_new['hour'][i] <= 11 and dataset_final_new['hour'][i] >= 6 :
#             dataset_final_new['status_day'][i] = "morning"
#         elif dataset_final_new['hour'][i] <= 17 and dataset_final_new['hour'][i] >= 12 :
#             dataset_final_new['status_day'][i] = "afternoon"
#         elif dataset_final_new['hour'][i] <= 23 and dataset_final_new['hour'][i] >= 18 :
#             dataset_final_new['status_day'][i] = "evening"
#         elif dataset_final_new['hour'][i] <= 0 and dataset_final_new['hour'][i] >= 5 :
#             dataset_final_new['status_day'][i] = "night"

            
#     return dataset_final_new

# status_of_day(10)

# for i in range(len(dataset_final_new)) :
#     dataset_final_new['status_day'][i] =  status_of_day(dataset_final_new['hour'][i])

            ***************************************************
            ***************************************************
            ***************************************************

compare_userloclist = pd.read_csv('D://A_thesis/gowalla/out_random/compare_userloclist.csv')

compare_userloclist['check-in time'] = compare_userloclist['check-in time'].map(lambda x: datetime.strptime(x, "%Y-%m-%d %H:%M:%S"))
compare_userloclist['time-24'] = pd.to_datetime(compare_userloclist['check-in time']).dt.strftime('%H:%M:%S')

compare_userloclist.dtypes

compare_userloclist['hour'] = compare_userloclist['check-in time'].apply(lambda x: x.hour)

            ***************************************************
            ***************************************************
            ***************************************************

compare_userloclist['status_day'] = "test"

            ***************************************************
            ***************************************************
            ***************************************************

compare_userloclist["DayOfWeek"] = compare_userloclist["check-in time"].dt.weekday

compare_userloclist["dayName"] = pd.to_datetime(compare_userloclist['check-in time']).dt.strftime("%A")
compare_userloclist["dayName"]

                    ***************************************************
                    ***************************************************
                    ***************************************************

compare_userloclist['hour'].min()

compare_userloclist_new = compare_userloclist

compare_userloclist_new = compare_userloclist_new.drop(columns = {'Unnamed: 0' , 'Unnamed: 0.1' , 'Unnamed: 0.1.1'})

compare_userloclist_new

def status_of_day(h):
    compare_userloclist_new['status_day'] = "dsgrtr"
    for i in range(len(compare_userloclist_new)) :
        if compare_userloclist_new['hour'][i] <= 11 and compare_userloclist_new['hour'][i] >= 6 :
            compare_userloclist_new['status_day'][i] = "morning"
        if compare_userloclist_new['hour'][i] <= 17 and compare_userloclist_new['hour'][i] >= 12 :
            compare_userloclist_new['status_day'][i] = "afternoon"
        if compare_userloclist_new['hour'][i] <= 23 and compare_userloclist_new['hour'][i] >= 18 :
            compare_userloclist_new['status_day'][i] = "evening"
        if compare_userloclist_new['hour'][i] <= 5 and compare_userloclist_new['hour'][i] >= 0 :
            compare_userloclist_new['status_day'][i] = "night"

            
    return compare_userloclist_new

status_of_day(10)

compare_userloclist_new.to_csv('D://A_thesis/gowalla/out_random/work_with_time.csv')

compare_userloclist_new = pd.read_csv('D://A_thesis/gowalla/out_random/work_with_time.csv')

# for i , row in compare_userloclist_new.iterrows() :
#     compare_userloclist_new['status_day'][i] = status_of_day(compare_userloclist_new['hour'][i])

# for i in range(len(compare_userloclist_new)) :
#     compare_userloclist_new['status_day'] =  status_of_day(compare_userloclist_new['hour'][i])

compare_userloclist_new['status_day']

                    ***************************************************
                    ***************************************************
                    ***************************************************

compare_userloclist_new = pd.read_csv('D://A_thesis/gowalla/out_random/work_with_time.csv')

del compare_userloclist_new['Unnamed: 0']

def status_of_day(h):
    compare_userloclist_new['second_mode'] = "dsgrtr"
    for i in range(len(compare_userloclist_new)) :
        if compare_userloclist_new['hour'][i] <= 11 and compare_userloclist_new['hour'][i] >= 6 :
            compare_userloclist_new['second_mode'][i] = "morning"
        if compare_userloclist_new['hour'][i] <= 17 and compare_userloclist_new['hour'][i] >= 12 :
            compare_userloclist_new['second_mode'][i] = "afternoon"
        if compare_userloclist_new['hour'][i] <= 21 and compare_userloclist_new['hour'][i] >= 18 :
            compare_userloclist_new['second_mode'][i] = "evening"
        if compare_userloclist_new['hour'][i] <= 24 and compare_userloclist_new['hour'][i] >= 22 :
            compare_userloclist_new['second_mode'][i] = "night"   
        if compare_userloclist_new['hour'][i] <= 5 and compare_userloclist_new['hour'][i] >= 0 :
            compare_userloclist_new['second_mode'][i] = "night"

            
    return compare_userloclist_new

status_of_day(22)

def status_of_day(h):
    compare_userloclist_new['third_mode'] = "dsgrtr"
    for i in range(len(compare_userloclist_new)) :
        if compare_userloclist_new['hour'][i] <= 3 and compare_userloclist_new['hour'][i] >= 0 :
            compare_userloclist_new['third_mode'][i] = "time_0-3"
        if compare_userloclist_new['hour'][i] <= 7 and compare_userloclist_new['hour'][i] >= 4 :
            compare_userloclist_new['third_mode'][i] = "time_4-7"
        if compare_userloclist_new['hour'][i] <= 11 and compare_userloclist_new['hour'][i] >= 8 :
            compare_userloclist_new['third_mode'][i] = "time_8-11"
        if compare_userloclist_new['hour'][i] <= 15 and compare_userloclist_new['hour'][i] >= 12 :
            compare_userloclist_new['third_mode'][i] = "time_12-15"   
        if compare_userloclist_new['hour'][i] <= 19 and compare_userloclist_new['hour'][i] >= 16 :
            compare_userloclist_new['third_mode'][i] = "time_16-19"
        if compare_userloclist_new['hour'][i] <= 23 and compare_userloclist_new['hour'][i] >= 20 :
            compare_userloclist_new['third_mode'][i] = "time_20-23"

            
    return compare_userloclist_new

status_of_day(22)

compare_userloclist_new

compare_userloclist_new.to_csv('D://A_thesis/gowalla/out_random/final_compare_userloclist.csv')

                    ***************************************************
                    ***************************************************
                    ***************************************************

# days = ["Monday", "Tuesday", "Wednesday","Thursday", "Friday", "Saturday", "Sunday"]
# d = days[dataset_final["DayOfWeek"].weekday()]
