==> 15_calculate_precission_friendship week.ipynb <==
import pandas as pd
import numpy as np

friendshipdf_final = pd.read_csv('D://A_thesis/gowalla/out_random/second_idea/friendship_df_without_threshold.csv')

del friendshipdf_final['Unnamed: 0']

friendshipdf_final

pre_rec = pd.DataFrame()

user_info = friendshipdf_final.loc[friendshipdf_final['userid'] == 1]

user_info

user_info['friend_list_new'].values[0]

friend_list_new = list(map(int,str(user_info['friend_list_new'].values[0]).split(',')))

# friend_list_new

fr_n5_new = list(map(int,str(user_info['fr_n5_first_method_new'].values[0]).split(',')))

eshterak_topn5_friendlist = set(friend_list_new).intersection(set(fr_n5_new))

eshterak_topn5_friendlist = list(eshterak_topn5_friendlist)
eshterak_topn5_friendlist

eshterak_topn5_friendlist = np.abs(eshterak_topn5_friendlist)

len(eshterak_topn5_friendlist)

precission_n5 = len(eshterak_topn5_friendlist) / len(fr_n5_new)
precission_n5

recall_n5 = len(eshterak_topn5_friendlist) / len(friend_list_new)
recall_n5

                                *****************************************
                                *****************************************
                                *****************************************

fr_n10_first_method_new = list(map(int,str(user_info['fr_n10_first_method_new'].values[0]).split(',')))
fr_n15_first_method_new = list(map(int,str(user_info['fr_n15_first_method_new'].values[0]).split(',')))
fr_n20_first_method_new = list(map(int,str(user_info['fr_n20_first_method_new'].values[0]).split(',')))
fr_n25_first_method_new = list(map(int,str(user_info['fr_n25_first_method_new'].values[0]).split(',')))
fr_n30_first_method_new = list(map(int,str(user_info['fr_n30_first_method_new'].values[0]).split(',')))
fr_n35_first_method_new = list(map(int,str(user_info['fr_n35_first_method_new'].values[0]).split(',')))
fr_n40_first_method_new = list(map(int,str(user_info['fr_n40_first_method_new'].values[0]).split(',')))

eshterak_topn10_friendlist = set(friend_list_new).intersection(set(fr_n10_first_method_new))

eshterak_topn10_friendlist = list(eshterak_topn10_friendlist)
eshterak_topn10_friendlist

eshterak_topn10_friendlist = np.abs(eshterak_topn10_friendlist)

precission_n10 = len(eshterak_topn10_friendlist) / len(fr_n10_first_method_new)
precission_n10

recall_n10 = len(eshterak_topn10_friendlist) / len(friend_list_new)
recall_n10

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn15_friendlist = set(friend_list_new).intersection(set(fr_n15_first_method_new))

eshterak_topn15_friendlist = set(friend_list_new).intersection(set(fr_n15_first_method_new))

eshterak_topn15_friendlist = list(eshterak_topn15_friendlist)
eshterak_topn15_friendlist

eshterak_topn15_friendlist = np.abs(eshterak_topn15_friendlist)

precission_n15 = len(eshterak_topn15_friendlist) / len(fr_n15_first_method_new)
precission_n15

recall_n15 = len(eshterak_topn15_friendlist) / len(friend_list_new)
recall_n15

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn20_friendlist = set(friend_list_new).intersection(set(fr_n20_first_method_new))

eshterak_topn20_friendlist = set(friend_list_new).intersection(set(fr_n20_first_method_new))

eshterak_topn20_friendlist = list(eshterak_topn20_friendlist)
eshterak_topn20_friendlist

eshterak_topn20_friendlist = np.abs(eshterak_topn20_friendlist)

precission_n20 = len(eshterak_topn20_friendlist) / len(fr_n20_first_method_new)
precission_n20

recall_n20 = len(eshterak_topn20_friendlist) / len(friend_list_new)
recall_n20

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn25_friendlist = set(friend_list_new).intersection(set(fr_n25_first_method_new))

eshterak_topn25_friendlist = set(friend_list_new).intersection(set(fr_n25_first_method_new))

eshterak_topn25_friendlist = list(eshterak_topn25_friendlist)
eshterak_topn25_friendlist

eshterak_topn25_friendlist = np.abs(eshterak_topn25_friendlist)

precission_n25 = len(eshterak_topn25_friendlist) / len(fr_n25_first_method_new)
precission_n25

recall_n25 = len(eshterak_topn25_friendlist) / len(friend_list_new)
recall_n25

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn30_friendlist = set(friend_list_new).intersection(set(fr_n30_first_method_new))

eshterak_topn30_friendlist = set(friend_list_new).intersection(set(fr_n30_first_method_new))

eshterak_topn30_friendlist = list(eshterak_topn30_friendlist)
eshterak_topn30_friendlist

eshterak_topn30_friendlist = np.abs(eshterak_topn30_friendlist)

precission_n30 = len(eshterak_topn30_friendlist) / len(fr_n30_first_method_new)
precission_n30

recall_n30 = len(eshterak_topn30_friendlist) / len(friend_list_new)
recall_n30

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn35_friendlist = set(friend_list_new).intersection(set(fr_n35_first_method_new))

eshterak_topn35_friendlist = set(friend_list_new).intersection(set(fr_n35_first_method_new))

eshterak_topn35_friendlist = list(eshterak_topn35_friendlist)
eshterak_topn35_friendlist

eshterak_topn35_friendlist = np.abs(eshterak_topn35_friendlist)

precission_n35 = len(eshterak_topn35_friendlist) / len(fr_n35_first_method_new)
precission_n35

recall_n35 = len(eshterak_topn35_friendlist) / len(friend_list_new)
recall_n35

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn40_friendlist = set(friend_list_new).intersection(set(fr_n40_first_method_new))

eshterak_topn40_friendlist = set(friend_list_new).intersection(set(fr_n40_first_method_new))

eshterak_topn40_friendlist = list(eshterak_topn40_friendlist)
eshterak_topn40_friendlist

eshterak_topn40_friendlist = np.abs(eshterak_topn40_friendlist)

precission_n40 = len(eshterak_topn40_friendlist) / len(fr_n40_first_method_new)
precission_n40

recall_n40 = len(eshterak_topn40_friendlist) / len(friend_list_new)
recall_n40

user_info['precission_n5'] = precission_n5
user_info['precission_n10'] = precission_n10
user_info['precission_n15'] = precission_n15
user_info['precission_n20'] = precission_n20

user_info['recall_n5']  = recall_n5
user_info['recall_n10']  = recall_n10
user_info['recall_n15']  = recall_n15
user_info['recall_n20']  = recall_n20

user_info['precission_n25'] = precission_n25
user_info['precission_n30'] = precission_n30
user_info['precission_n35'] = precission_n35
user_info['precission_n40'] = precission_n40

user_info['recall_n25']  = recall_n25
user_info['recall_n30']  = recall_n30
user_info['recall_n35']  = recall_n35
user_info['recall_n40']  = recall_n40

# user_info = user_info.drop(columns=['fr_n5', 'fr_n10' , 'fr_n15' , 'fr_n20'])

user_info

# user_info.to_csv('D://A_thesis/gowalla/out_random/precission_recall.csv')

                                *****************************************
                                *****************************************
                                *****************************************

                                *****************************************
                                *****************************************
                                *****************************************

                                  _______________second_mode_______________

                                *****************************************
                                *****************************************
                                *****************************************

                                *****************************************
                                *****************************************
                                *****************************************

fr_n5_second_method_new = list(map(int,str(user_info['fr_n5_second_method_new'].values[0]).split(',')))
fr_n10_second_method_new = list(map(int,str(user_info['fr_n10_second_method_new'].values[0]).split(',')))
fr_n15_second_method_new = list(map(int,str(user_info['fr_n15_second_method_new'].values[0]).split(',')))
fr_n20_second_method_new = list(map(int,str(user_info['fr_n20_second_method_new'].values[0]).split(',')))
fr_n25_second_method_new = list(map(int,str(user_info['fr_n25_second_method_new'].values[0]).split(',')))
fr_n30_second_method_new = list(map(int,str(user_info['fr_n30_second_method_new'].values[0]).split(',')))
fr_n35_second_method_new = list(map(int,str(user_info['fr_n35_second_method_new'].values[0]).split(',')))
fr_n40_second_method_new = list(map(int,str(user_info['fr_n40_second_method_new'].values[0]).split(',')))

eshterak_topn5_friendlist_second_method = set(friend_list_new).intersection(set(fr_n5_second_method_new))

eshterak_topn5_friendlist_second_method = list(eshterak_topn5_friendlist_second_method)
eshterak_topn5_friendlist_second_method

eshterak_topn5_friendlist_second_method = np.abs(eshterak_topn5_friendlist_second_method)

precission_n5_second_method = len(eshterak_topn5_friendlist_second_method) / len(fr_n5_second_method_new)
precission_n5_second_method

recall_n5_second_method = len(eshterak_topn5_friendlist_second_method) / len(friend_list_new)
recall_n5_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn10_friendlist_second_method = set(friend_list_new).intersection(set(fr_n10_second_method_new))

eshterak_topn10_friendlist_second_method = list(eshterak_topn10_friendlist_second_method)
eshterak_topn10_friendlist_second_method

eshterak_topn10_friendlist_second_method = np.abs(eshterak_topn10_friendlist_second_method)

precission_n10_second_method = len(eshterak_topn10_friendlist_second_method) / len(fr_n10_second_method_new)
precission_n10_second_method

recall_n10_second_method = len(eshterak_topn10_friendlist_second_method) / len(friend_list_new)
recall_n10_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn15_friendlist_second_method = set(friend_list_new).intersection(set(fr_n15_second_method_new))

eshterak_topn15_friendlist_second_method = list(eshterak_topn15_friendlist_second_method)
eshterak_topn15_friendlist_second_method

eshterak_topn15_friendlist_second_method = np.abs(eshterak_topn15_friendlist_second_method)

precission_n15_second_method = len(eshterak_topn15_friendlist_second_method) / len(fr_n15_second_method_new)
precission_n15_second_method

recall_n15_second_method = len(eshterak_topn15_friendlist_second_method) / len(friend_list_new)
recall_n15_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn20_friendlist_second_method = set(friend_list_new).intersection(set(fr_n20_second_method_new))

eshterak_topn20_friendlist_second_method = list(eshterak_topn20_friendlist_second_method)
eshterak_topn20_friendlist_second_method

eshterak_topn20_friendlist_second_method = np.abs(eshterak_topn20_friendlist_second_method)

precission_n20_second_method = len(eshterak_topn20_friendlist_second_method) / len(fr_n20_second_method_new)
precission_n20_second_method

recall_n20_second_method = len(eshterak_topn20_friendlist_second_method) / len(friend_list_new)
recall_n20_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn25_friendlist_second_method = set(friend_list_new).intersection(set(fr_n25_second_method_new))

eshterak_topn25_friendlist_second_method = list(eshterak_topn25_friendlist_second_method)
eshterak_topn25_friendlist_second_method

eshterak_topn25_friendlist_second_method = np.abs(eshterak_topn25_friendlist_second_method)

precission_n25_second_method = len(eshterak_topn25_friendlist_second_method) / len(fr_n25_second_method_new)
precission_n25_second_method

recall_n25_second_method = len(eshterak_topn25_friendlist_second_method) / len(friend_list_new)
recall_n25_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn30_friendlist_second_method = set(friend_list_new).intersection(set(fr_n30_second_method_new))

eshterak_topn30_friendlist_second_method = list(eshterak_topn30_friendlist_second_method)
eshterak_topn30_friendlist_second_method

eshterak_topn30_friendlist_second_method = np.abs(eshterak_topn30_friendlist_second_method)

precission_n30_second_method = len(eshterak_topn30_friendlist_second_method) / len(fr_n30_second_method_new)
precission_n30_second_method

recall_n30_second_method = len(eshterak_topn30_friendlist_second_method) / len(friend_list_new)
recall_n30_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn35_friendlist_second_method = set(friend_list_new).intersection(set(fr_n35_second_method_new))

eshterak_topn35_friendlist_second_method = list(eshterak_topn35_friendlist_second_method)
eshterak_topn35_friendlist_second_method

eshterak_topn35_friendlist_second_method = np.abs(eshterak_topn35_friendlist_second_method)

precission_n35_second_method = len(eshterak_topn35_friendlist_second_method) / len(fr_n35_second_method_new)
precission_n35_second_method

recall_n35_second_method = len(eshterak_topn35_friendlist_second_method) / len(friend_list_new)
recall_n35_second_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn40_friendlist_second_method = set(friend_list_new).intersection(set(fr_n40_second_method_new))

eshterak_topn40_friendlist_second_method = list(eshterak_topn40_friendlist_second_method)
eshterak_topn40_friendlist_second_method

eshterak_topn40_friendlist_second_method = np.abs(eshterak_topn40_friendlist_second_method)

precission_n40_second_method = len(eshterak_topn40_friendlist_second_method) / len(fr_n40_second_method_new)
precission_n40_second_method

recall_n40_second_method = len(eshterak_topn40_friendlist_second_method) / len(friend_list_new)
recall_n40_second_method

user_info['precission_n5_second_method'] = precission_n5_second_method
user_info['precission_n10_second_method'] = precission_n10_second_method
user_info['precission_n15_second_method'] = precission_n15_second_method
user_info['precission_n20_second_method'] = precission_n20_second_method
user_info['precission_n25_second_method'] = precission_n25_second_method
user_info['precission_n30_second_method'] = precission_n30_second_method
user_info['precission_n35_second_method'] = precission_n35_second_method
user_info['precission_n40_second_method'] = precission_n40_second_method

user_info['recall_n5_second_method']  = recall_n5_second_method
user_info['recall_n10_second_method']  = recall_n10_second_method
user_info['recall_n15_second_method']  = recall_n15_second_method
user_info['recall_n20_second_method']  = recall_n20_second_method
user_info['recall_n25_second_method'] = recall_n25_second_method
user_info['recall_n30_second_method'] = recall_n30_second_method
user_info['recall_n35_second_method'] = recall_n35_second_method
user_info['recall_n40_second_method'] = recall_n40_second_method

user_info

                                *****************************************
                                *****************************************
                                *****************************************

                                *****************************************
                                *****************************************
                                *****************************************

                                  _______________third_mode_______________

                                *****************************************
                                *****************************************
                                *****************************************

                                *****************************************
                                *****************************************
                                *****************************************

fr_n5_third_method_new = list(map(int,str(user_info['fr_n5_third_method_new'].values[0]).split(',')))
fr_n10_third_method_new = list(map(int,str(user_info['fr_n10_third_method_new'].values[0]).split(',')))
fr_n15_third_method_new = list(map(int,str(user_info['fr_n15_third_method_new'].values[0]).split(',')))
fr_n20_third_method_new = list(map(int,str(user_info['fr_n20_third_method_new'].values[0]).split(',')))
fr_n25_third_method_new = list(map(int,str(user_info['fr_n25_third_method_new'].values[0]).split(',')))
fr_n30_third_method_new = list(map(int,str(user_info['fr_n30_third_method_new'].values[0]).split(',')))
fr_n35_third_method_new = list(map(int,str(user_info['fr_n35_third_method_new'].values[0]).split(',')))
fr_n40_third_method_new = list(map(int,str(user_info['fr_n40_third_method_new'].values[0]).split(',')))

eshterak_topn5_friendlist_third_method = set(friend_list_new).intersection(set(fr_n5_third_method_new))

eshterak_topn5_friendlist_third_method = list(eshterak_topn5_friendlist_third_method)
eshterak_topn5_friendlist_third_method

eshterak_topn5_friendlist_third_method = np.abs(eshterak_topn5_friendlist_third_method)

precission_n5_third_method = len(eshterak_topn5_friendlist_third_method) / len(fr_n5_second_method_new)
precission_n5_third_method

recall_n5_third_method = len(eshterak_topn5_friendlist_third_method) / len(friend_list_new)
recall_n5_third_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn10_friendlist_third_method = set(friend_list_new).intersection(set(fr_n10_third_method_new))

eshterak_topn10_friendlist_third_method = list(eshterak_topn10_friendlist_third_method)
eshterak_topn10_friendlist_third_method

eshterak_topn10_friendlist_third_method = np.abs(eshterak_topn10_friendlist_third_method)

precission_n10_third_method = len(eshterak_topn10_friendlist_third_method) / len(fr_n10_second_method_new)
precission_n10_third_method

recall_n10_third_method = len(eshterak_topn10_friendlist_third_method) / len(friend_list_new)
recall_n10_third_method


                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn15_friendlist_third_method = set(friend_list_new).intersection(set(fr_n15_third_method_new))

eshterak_topn15_friendlist_third_method = list(eshterak_topn15_friendlist_third_method)
eshterak_topn15_friendlist_third_method

eshterak_topn15_friendlist_third_method = np.abs(eshterak_topn15_friendlist_third_method)

precission_n15_third_method = len(eshterak_topn15_friendlist_third_method) / len(fr_n15_second_method_new)
precission_n15_third_method

recall_n15_third_method = len(eshterak_topn15_friendlist_third_method) / len(friend_list_new)
recall_n15_third_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn20_friendlist_third_method = set(friend_list_new).intersection(set(fr_n20_third_method_new))

eshterak_topn20_friendlist_third_method = list(eshterak_topn20_friendlist_third_method)
eshterak_topn20_friendlist_third_method

eshterak_topn20_friendlist_third_method = np.abs(eshterak_topn20_friendlist_third_method)

precission_n20_third_method = len(eshterak_topn20_friendlist_third_method) / len(fr_n20_second_method_new)
precission_n20_third_method

recall_n20_third_method = len(eshterak_topn20_friendlist_third_method) / len(friend_list_new)
recall_n20_third_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn25_friendlist_third_method = set(friend_list_new).intersection(set(fr_n25_third_method_new))

eshterak_topn25_friendlist_third_method = list(eshterak_topn25_friendlist_third_method)
eshterak_topn25_friendlist_third_method

eshterak_topn25_friendlist_third_method = np.abs(eshterak_topn25_friendlist_third_method)

precission_n25_third_method = len(eshterak_topn25_friendlist_third_method) / len(fr_n25_second_method_new)
precission_n25_third_method

recall_n25_third_method = len(eshterak_topn25_friendlist_third_method) / len(friend_list_new)
recall_n25_third_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn30_friendlist_third_method = set(friend_list_new).intersection(set(fr_n30_third_method_new))

eshterak_topn30_friendlist_third_method = list(eshterak_topn30_friendlist_third_method)
eshterak_topn30_friendlist_third_method

eshterak_topn30_friendlist_third_method = np.abs(eshterak_topn30_friendlist_third_method)

precission_n30_third_method = len(eshterak_topn30_friendlist_third_method) / len(fr_n30_second_method_new)
precission_n30_third_method

recall_n30_third_method = len(eshterak_topn30_friendlist_third_method) / len(friend_list_new)
recall_n30_third_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn35_friendlist_third_method = set(friend_list_new).intersection(set(fr_n35_third_method_new))

eshterak_topn35_friendlist_third_method = list(eshterak_topn35_friendlist_third_method)
eshterak_topn35_friendlist_third_method

eshterak_topn35_friendlist_third_method = np.abs(eshterak_topn35_friendlist_third_method)

precission_n35_third_method = len(eshterak_topn35_friendlist_third_method) / len(fr_n35_second_method_new)
precission_n35_third_method

recall_n35_third_method = len(eshterak_topn35_friendlist_third_method) / len(friend_list_new)
recall_n35_third_method

                                *****************************************
                                *****************************************
                                *****************************************

eshterak_topn40_friendlist_third_method = set(friend_list_new).intersection(set(fr_n40_third_method_new))

eshterak_topn40_friendlist_third_method = list(eshterak_topn40_friendlist_third_method)
eshterak_topn40_friendlist_third_method

eshterak_topn40_friendlist_third_method = np.abs(eshterak_topn40_friendlist_third_method)

precission_n40_third_method = len(eshterak_topn40_friendlist_third_method) / len(fr_n40_second_method_new)
precission_n40_third_method

recall_n40_third_method = len(eshterak_topn40_friendlist_third_method) / len(friend_list_new)
recall_n40_third_method

                                *****************************************
                                *****************************************
                                *****************************************

user_info['precission_n5_third_method'] = precission_n5_third_method
user_info['precission_n10_third_method'] = precission_n10_third_method
user_info['precission_n15_third_method'] = precission_n15_third_method
user_info['precission_n20_third_method'] = precission_n20_third_method
user_info['precission_n25_third_method'] = precission_n25_third_method
user_info['precission_n30_third_method'] = precission_n30_third_method
user_info['precission_n35_third_method'] = precission_n35_third_method
user_info['precission_n40_third_method'] = precission_n40_third_method

user_info['recall_n5_third_method']  = recall_n5_third_method
user_info['recall_n10_third_method']  = recall_n10_third_method
user_info['recall_n15_third_method']  = recall_n15_third_method
user_info['recall_n20_third_method']  = recall_n20_third_method
user_info['recall_n25_third_method'] = recall_n25_third_method
user_info['recall_n30_third_method'] = recall_n30_third_method
user_info['recall_n35_third_method'] = recall_n35_third_method
user_info['recall_n40_third_method'] = recall_n40_third_method

user_info

user_info.to_csv('D://A_thesis/gowalla/out_random/second_idea/precission_recall_weekday_weekend.csv')

                                *****************************************
                                *****************************************
                                *****************************************
